Update Readme & utils.go creation
- update the Readme
- create utils.go file
