+++ README.md
@@ -1,6 +1,6 @@
 # Cosmos Scaffold
 
-This is a scaffolding tool for CosmosSDK based applications. To build the binary, have v1.13 [golang installed](https://golang.org/doc/install) and then just run:
+This is a scaffolding tool for CosmosSDK based applications. To build the binary, have v1.13 [golang installed](https://golang.org/doc/install) and then run:
 
 ```bash
 make
@@ -12,11 +12,13 @@ Then the binary should be installed in your `$GOBIN` so you can then run:
 scaffold --help
 ```
 
+## Tutorial
+
 To scaffold out the `nameservice` tutorial example just run the following:
 
 ```bash
 cd $GOPATH/src/github.com/{{ .Username }}
-scaffold nameservice myghusername myrepo
+scaffold tutorial nameservice myghusername myrepo
 ```
 
 Then just `cd` into the `myrepo` folder and:
@@ -27,3 +29,11 @@ make install
 nscli --help
 nsd --help
 ```
+
+# Versioning
+
+This repo will have its own versioning and contain a compatibility table:
+
+| Cosmos-SDK Version | Scaffold Version |
+| ------------------ | :--------------: |
+| 0.37.\*            |      0.1.0       |
